simplify handling of themes and make reload quicker; render console only on client side (for now)
